tests/wpt/tests/html/interaction/focus/focusgroup/tentative/sequential-navigation/directional-input-handler-only-item.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>HTML Test: focusgroup - Native directional input handler as only item in segment</title>
<link rel="author" title="Microsoft" href="http://www.microsoft.com/">
<link rel="help" href="https://open-ui.org/components/scoped-focusgroup.explainer/">
<meta name="assert" content="When a native directional input handler is the only item in a segment, the segment still has an entry point when the element is not focused.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/resources/testdriver-actions.js"></script>
<script src="/shadow-dom/focus-navigation/resources/focus-utils.js"></script>
<script src="../resources/focusgroup-utils.js"></script>

<button tabindex=0 id="before">Before</button>
<div id="toolbar" focusgroup="toolbar">
  <input id="search" type="search" placeholder="Search">
</div>
<button tabindex=0 id="after">After</button>

<script>

  promise_test(async t => {
    const search = document.getElementById("search");

    await focusAndWait(search);

    // Directional input should not move focus (native directional input handling).
    await assert_directional_input_does_not_move_focus(search);
  }, "Directional input does not work from native directional input handler even when it's the only item");

  promise_test(async t => {
    const before = document.getElementById("before");
    const search = document.getElementById("search");
    const after = document.getElementById("after");

    // Tab navigation: enter at search, exit to after.
    await assert_focusgroup_tab_navigation([before, search, after]);

    // Shift+Tab from after returns to focused search.
    await navigateFocusBackward();
    assert_equals(document.activeElement, search);
  }, "Segment with only native directional input handler navigates correctly based on focus state");

</script>
